enhance: bake to archival dir
