+++ baker/GrapherBaker.tsx
@@ -329,7 +329,7 @@ export const bakeSingleGrapherPageForArchival = async (
         outPathManifest,
         JSON.stringify({ staticAssetMap, runtimeAssetMap }, undefined, 2)
     )
-    console.log(outPathHtml, outPathManifest)
+    console.log(outPathHtml)
 }
 
 const bakeGrapherPage = async (
